requirements.txt
numpy
pandas
scikit-learn
transformers
sentence-transformers
faiss-cpu
lucene
pyserini
tqdm
